qase-cucumberjs: add a storage
Move parsing logic to `storage` class from the reporter.
